% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.support-de.R
\name{lang.support.de}
\alias{lang.support.de}
\title{Language support for German}
\usage{
lang.support.de(...)
}
\arguments{
\item{...}{Optional arguments for \code{\link[koRpus:set.lang.support]{set.lang.support}}.}
}
\description{
This function adds support for German to the koRpus package. You should not
need to call it manually, as that is done automatically when this package is
being loaded.
}
\details{
In particular, this function adds the following:
\itemize{
 \item \code{lang}: The additional language "de" to be used with koRpus
 \item \code{treetag}: The additional preset "de",
      implemented according to the respective TreeTagger[1] script
 \item \code{POS tags}: An additional set of tags,
      implemented using the documentation for the corresponding
   TreeTagger parameter set[2]
}
Hyphenation patterns are provided by means of the \code{\link[sylly.de:hyph.support.de]{sylly.de}} package.
}
\examples{
lang.support.de()
}
\references{
[1] \url{http://www.cis.uni-muenchen.de/~schmid/tools/TreeTagger/}

[2] \url{http://www.cis.uni-muenchen.de/~schmid/tools/TreeTagger/data/stts_guide.pdf}
}
